backend/requirements.txt
django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
django-redis
redis
python-dotenv
cryptography
django-otp
qrcode